+++ src/App.tsx
@@ -14,11 +14,7 @@ function App() {
           <Routes>
             <Route path="/login" element={<LoginPage />} />
             <Route path="/signup" element={<SignupPage />} />
-            <Route path="*" element={
-              <ProtectedRoute>
-                <AppShell />
-              </ProtectedRoute>
-            } />
+            <Route path="*" element={<AppShell />} />
           </Routes>
         </MotionProvider>
       </AuthProvider>